backend/requirements.txt
absl-py==2.1.0
accelerate==1.8.1
asgiref==3.9.1
astunparse==1.6.3
blinker==1.9.0
certifi==2024.12.14
cffi==2.0.0
channels==4.2.2
charset-normalizer==3.4.1
click==8.2.1
colorama==0.4.6
cryptography==46.0.1
Django==5.2.6
django-cors-headers==4.8.0
djangorestframework==3.16.1
filelock==3.18.0
Flask==3.1.1
flatbuffers==24.12.23
fsspec==2025.5.1
gast==0.6.0
google-pasta==0.2.0
grpcio==1.69.0
h5py==3.12.1
huggingface-hub==0.33.1
ibm-cos-sdk==2.13.6
ibm-cos-sdk-core==2.13.6
ibm-cos-sdk-s3transfer==2.13.6
ibm_watson_machine_learning==1.0.367
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.4.2
keras==3.8.0
libclang==18.1.1
lomond==0.3.3
lxml==6.0.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml-dtypes==0.4.1
mpmath==1.3.0
mysqlclient==2.2.7
namex==0.0.8
narwhals==1.44.0
networkx==3.5
numpy==1.26.4
opt_einsum==3.4.0
optree==0.13.1
packaging==24.2
pandas==2.1.4
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.1.0
plotly==6.2.0
protobuf==5.29.3
psutil==7.0.0
pycparser==2.23
Pygments==2.19.1
PyPDF2==3.0.1
pypdfium2==4.30.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.2.0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.2
rich==13.9.4
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.1
setuptools==75.8.0
six==1.17.0
sqlparse==0.5.3
sympy==1.14.0
tabulate==0.9.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow_intel==2.18.0
termcolor==2.5.0
threadpoolctl==3.5.0
tokenizers==0.21.2
torch==2.7.1
torchaudio==2.7.1
torchvision==0.22.1
tqdm==4.67.1
transformers==4.53.0
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.3.0
waitress==3.0.2
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.1
zipp==3.23.0
PyPDF2>=3.0.0
pdfplumber==0.10.0
python-docx>=0.8.11
pytesseract==0.3.10
Pillow==10.0.0
spacy>=3.7.0
nltk==3.8.1
